+++ plugins/svg.js
@@ -1,4 +1,3 @@
-import textLoader from "../loaders/text.js";
 import { optimize } from "../deps/svgo.js";
 import { merge } from "../utils.js";
 
@@ -12,7 +11,7 @@ export default function (userOptions) {
   const options = merge(defaults, userOptions);
 
   return (site) => {
-    site.loadAssets(options.extensions, textLoader);
+    site.loadAssets(options.extensions);
     site.process(options.extensions, processor);
 
     async function processor(page) {