+++ CHANGELOG.md
@@ -7,6 +7,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## Unreleased
+### Fixed
+- Updated `std`.
+
 ## [0.20.2] - 2021-05-18
 ### Added
 - More default character replacements to slugifier.
@@ -17,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
   This adds support for more features, like pipes, etc.
 
 ### Fixed
-- Updated `pug` and `std`.
+- Updated `pug`.
 - Resolve the imported files from template engines, like `nunjucks`,
   when the `src` is in a subdirectory.
 
@@ -719,6 +723,7 @@ The first version.
 [#94]: https://github.com/lumeland/lume/issues/94
 [#95]: https://github.com/lumeland/lume/issues/95
 
+[Unreleased]: https://github.com/lumeland/lume/compare/v0.20.2...HEAD
 [0.20.2]: https://github.com/lumeland/lume/compare/v0.20.1...v0.20.2
 [0.20.1]: https://github.com/lumeland/lume/compare/v0.20.0...v0.20.1
 [0.20.0]: https://github.com/lumeland/lume/compare/v0.19.0...v0.20.0