+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [Unreleased]
+## [0.21.0] - 2021-06-01
 ### Added
 - The `postcss` and `terser` filters.
 - The `--plugins` argument to `lume init`, so you can load and use plugins.
@@ -755,7 +755,7 @@ The first version.
 [#94]: https://github.com/lumeland/lume/issues/94
 [#95]: https://github.com/lumeland/lume/issues/95
 
-[Unreleased]: https://github.com/lumeland/lume/compare/v0.20.2...HEAD
+[0.21.0]: https://github.com/lumeland/lume/compare/v0.20.2...v0.21.0
 [0.20.2]: https://github.com/lumeland/lume/compare/v0.20.1...v0.20.2
 [0.20.1]: https://github.com/lumeland/lume/compare/v0.20.0...v0.20.1
 [0.20.0]: https://github.com/lumeland/lume/compare/v0.19.0...v0.20.0